Rename of dicussion-link.js to discussion-link.js
